requirements.txt
django==3.1.6
djangorestframework==3.12.2